# fastdelta: remove sort allocations