While looking at the internal testing that's going on, I noticed a lot of allocations caused by a `sort.Slice()`.

<img width="1880" alt="image" src="https://user-images.githubusercontent.com/15000/195282827-3971f440-0d02-473e-bcd7-faf6d60d1bda.png">

I think we can easily get rid of them by applying this PR. The benchmarks says this will reduce allocated objects by ~60% and CPU time by 10%.

Let me know what you think.

```
name                                  old time/op    new time/op    delta
FastDelta/testdata/heap.pprof-12         924µs ± 2%     827µs ± 3%  -10.53%  (p=0.008 n=5+5)
FastDelta/testdata/big-heap.pprof-12    12.2ms ± 4%    10.9ms ± 4%  -10.29%  (p=0.008 n=5+5)

name                                  old alloc/op   new alloc/op   delta
FastDelta/testdata/heap.pprof-12         289kB ± 0%     226kB ± 0%  -22.00%  (p=0.008 n=5+5)
FastDelta/testdata/big-heap.pprof-12    2.44MB ± 0%    1.71MB ± 0%  -29.77%  (p=0.008 n=5+5)

name                                  old allocs/op  new allocs/op  delta
FastDelta/testdata/heap.pprof-12         3.33k ± 0%     1.34k ± 0%  -59.66%  (p=0.008 n=5+5)
FastDelta/testdata/big-heap.pprof-12     37.2k ± 0%     14.5k ± 0%  -61.10%  (p=0.008 n=5+5)
```